+++ vantage6-algorithm-tools/vantage6/algorithm/mock/__init__.py
@@ -1,3 +1,3 @@
-from vantage6.mock.network import MockNetwork
+from vantage6.algorithm.mock.network import MockNetwork
 
 __all__ = ["MockNetwork"]
